Re-work when original_font_checksum, original_features, original_axis_space, and codepoint_ordering are sent.
Change to have these font derived fields sent whenever the server sees a mismatch between the client reported original font checksum and the servers original font checksum. This allows the fields to be updated in the client's state during a font version upgrade. Such as when a server is patching a client to a new version of the original font.