Revamp the FeatureTagSet encoding scheme.
'default included' features can now be omitted from the set by including the ID associated with that feature. This allows the client to fully customize the feature set if the list of default included features does not match what the client's shaper is expected to require for it's current content. For example the 'vert' feature could be omitted if no vertical layout is being done.