Update references to RFC9110, RFC9111 and RFC9112
We eventually switched to the `https://httpwg.org/specs/` versions for RFC9110,
RFC9111 and RFC9112 in the cross-references database, because they are more
readable.

Unfortunately, this means that references to sections in these specs need to be
updated again, as was done in #144. This update uses the new (and in theory now
stable) fragment identifiers.