Remove conformance keyword from 'font subset' definition. (#295)
This definition is just describing what the concept of a font subset is and should not be including conformance requirements. It's better to state those more directly in the relevant sections later on. In this case requirements for font subset rendering equivalence in encodings are currently introduced in the encoder section. Fixes #288